Separate API interface for removing source item configuration
